server: Added error messages

+++ server/routes/file-folder-crud.js
@@ -7,6 +7,13 @@ const fs = require("fs");
 router.post("/rm", (req, res) => {
 	const body = req.body;
 	const path = `${__dirname}/../uploads${body.path}`;
+
+	// Check if file exists
+	if (!fs.existsSync(path)) {
+		res.json({ err: true, errMsg: "File you want to delete does not exists." });
+		return;
+	}
+
 	console.log("File Deleted : ", path);
 	if (body.isFolder) {
 		fs.rmSync(path, { recursive: true, force: true });
@@ -21,7 +28,14 @@ router.post("/rm", (req, res) => {
 router.post("/upload", (req, res) => {
 	const body = req.body;
 	const file = req.files.file;
-	file.mv(`${__dirname}/../uploads/${body.path}/${file.name}`);
+
+	try {
+		file.mv(`${__dirname}/../uploads/${body.path}/${file.name}`);
+	} catch (err) {
+		res.json({ err: true, errMsg: "Upload Failed." });
+		return;
+	}
+
 	console.log("File Uploaded : ", file.name);
 	res.end();
 });
@@ -31,7 +45,10 @@ router.post("/mkdir", (req, res) => {
 	const body = req.body;
 	const folder = `${__dirname}/../uploads/${body.name}`;
 	console.log(folder);
-	if (!fs.existsSync(folder)) {
+	if (fs.existsSync(folder)) {
+		res.json({ err: true, errMsg: "Folder allready exists" });
+		return;
+	} else {
 		fs.mkdirSync(folder);
 	}
 	res.end();
